Refactor ReceiverComponent into background package.
