+++ app/src/main/java/io/homeassistant/companion/android/background/LocationBroadcastReceiver.kt
@@ -14,7 +14,6 @@ import androidx.core.app.ActivityCompat
 import com.google.android.gms.location.LocationRequest
 import com.google.android.gms.location.LocationResult
 import com.google.android.gms.location.LocationServices
-import io.homeassistant.companion.android.DaggerReceiverComponent
 import io.homeassistant.companion.android.common.dagger.Graph
 import io.homeassistant.companion.android.domain.integration.IntegrationUseCase
 import io.homeassistant.companion.android.domain.integration.UpdateLocation
