+++ app/src/main/java/io/homeassistant/companion/android/background/ReceiverComponent.kt
@@ -1,7 +1,6 @@
-package io.homeassistant.companion.android
+package io.homeassistant.companion.android.background
 
 import dagger.Component
-import io.homeassistant.companion.android.background.LocationBroadcastReceiver
 import io.homeassistant.companion.android.common.dagger.AppComponent
 
 @Component(dependencies = [AppComponent::class])
